sos: Pluggable indexes for a SOS store
This patch implements a pluggable index layer and modifies SOS to
use this index abstraction. A B+ Tree index was implemented as
the first index plug-in. The abstraction also supports pluggable
key-comparators so that arbitrary user data types can be
supported. Initial support is provided for uint32/64 and string.

This change also modify other programs to use the new SOS.

Change-Id: Id36d00eabd2f1a5d81f0f820c5fe83e777be9001
Signed-off-by: Tom Tucker <tom@ogc.us>