Add connection data to Zap connection setup
When a connection is established between two peers, the
active side provides a data buffer with connection data that
is provided to the peer in the connection request event. If the
peer accepts the connection, it may also send connection
data back to the active side that is provided in the connected event.

This intended use for this feature is to allow applications to exchange
user authentication data, and connection configuration data.

Change-Id: I7d6a644163e3a9ff51177f2cafb7257bf2dca773
Signed-off-by: Tom Tucker <tom@ogc.us>