baler: bquery timestamp key bug fix
When moving from SOS2 to SOS3 we forgot to change query key in message
query case. The bug causes 'bquery -B XXX' to not work due to wrong
key value set (should be 8 bytes, but set 4 bytes--and we're setting
the wrong 4 bytes). This commit corrects bquery to use union
sos_timestamp_u as the key.

Change-Id: Ia0a7d1d8efa4bd003b3b3506389834dd8f46028f