baler: handle missing pattern in bhttpd_query
The slave balerds will have only patterns it extracted. The master
balerd may have all patterns, but may not have complete statistics on
the patterns--because the statistics on patterns are updated based on
the messages (each message will only be processed by one of the
balerds). This commit update bhttpd_query so that it is more forgiving
on some missing pattern information.

NOTE: For multiple balerds scheme, the client will have to query all
bhttpds and aggregate the results by itself.

Change-Id: If4cd957954ae9f1a77b0fd5abb94e7bacf594cfb