Resolve -Werror -Wall warnings in ldms
Turned on -Wall and -Werror and resolved all warnings in ldms source code. This
makes it feasible for developers to turn this on as a matter of course and have a
a policy whereby checking in warnings is discouraged.

As part of this, exception handing is added to the ldms swig interface so that type
errors will generate a Python TypError Exception

Signed-off-by: Tom Tucker <tom@ogc.us>