rhel7: lib,ldms,baler - installation scriptlets
Add post-installation, pre-uninstallation, and post-uninstallation
scriptlets for lib, ldms and baler.

For lib, post-installation creates a soft-link
/etc/ld.so.conf.d/ovis-ld-so.conf, and run ldconfig so that the
dynamic loader recognizes the installed library. post-uninstallation
for lib removes the softlink and invoke ldconfig to update the loader.

For ldms and baler, post-installation calls ldconfig to update the
loader. The scriplet also creates systemd unit files and call
systemctl so that the system recognizes the new services. They are not
`enabled`. For pre-uninstallation, the scriptlet stop the services.
Then, in post-uninstallation scriptlet, the unit files are removed,
and systemctl and ldconfig are called to notify the removal of
services and libraries.