baler: Bq.Bstore.open() no-create by default
Add `flags` and `mode` parameters to `Bq.Bstore.open()` method. Also,
if `flags` and `mode` are not given, the default are `O_RDWR`, and
`0660` respectively.

Existing python applications that intend to open, not to create,
Bstore should not be affected by this change, given that the store is
created by `balerd` first.