baler: bstore_sos - long token segfault bug fix
bstore_sos.tkn_add() uses sos_key on the stack, which does not support
a token longer than 256 octets. tkn_add() got segmentation fault when
encountered a long token, as it overflow the stack key. This patch
fixes this problem by dynamically allocating the new key if the token
length is too long.