baler: bin_tcp - add max_msg_len parameter
Add `max_msg_len` parameter to the `bin_tcp` input plugin, to limit
the message length to be processed by the plugin. The longer message
will be truncated. The default value is 0, which is unlimited.

TEST: The truncation capability has been manually tested (set
BTEST_CHECK='' and BTEST_BIN_MAX_MSG_LEN=65) as the automatic checking
script does not yet have the capability to take truncation yet (as it
also affect patterns).