bstore_sos - set_pos() segfault bug fix
Considering the following use case:

```
itr0 = iter_new()
iter_begin(itr0)
iter_next(itr0)
iter_next(itr0)
pos0 = iter_pos_get(itr0)
itr1 = iter_new()
iter_pos_set(itr1, pos0)
```

The pos_set() results in segmentation fault for msg_iter in bstore_sos
because the internal iterator has not been initialized yet due to the
fact that it doesn't know which index to use at the time.

This patch fix the bug by checking if the internal sos iterator has
been initialized. If not, it will initialize the iterator according to
the type encoded in the bstore_iter_pos_t.

This patch also update the simple_test4/check-scripts/check_bstore.py
so that the iterators in the position test cases are initialized with
the same parameters (e.g. time range), as positions are not meant to
be used to store the whole state of the iterator.