git info - get git info only from baler tree
Git information obtained by configure script could be incorrect in the
case of installing baler from a tar ball from within another git
repository. An example of this is making baler RPM from an RPM repo
with plain baler tar ball (not as a submodule). This patch fix the bug
by obtaining the SHA from git only if the source dir is a working
baler git tree. Otherwise, use the provided SHA.txt and TAG.txt (which
are from a baler working tree).