Improve test_util.BalerDaemon start/stop sequence
`test_util.BalerDaemon.start()` now will wait for "Baler is ready"
message in the log before returning. This is to make sure that the
daemon is ready to process the data.

This patch also add `test_util.BalerDaemon.wait_idle()` so that the
caller can blocked wait until the daemon is idle (a sign of finished
processing the data).